REFACTOR: Replacing specific filter by skill functions with generic one for all material children
